relay_handler: break lazy code out into another file
